# search: match repos concurrently