--- a/cmd/frontend/graphqlbackend/search_repositories_test.go
+++ b/cmd/frontend/graphqlbackend/search_repositories_test.go
@@ -6,11 +6,13 @@ import (
 	"reflect"
 	"regexp"
 	"sort"
+	"strconv"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
+	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/search"
 	searchbackend "github.com/sourcegraph/sourcegraph/internal/search/backend"
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
@@ -253,5 +255,35 @@ func TestMatchRepos(t *testing.T) {
 	}
 	if !reflect.DeepEqual(toMap(repos), toMap(want)) {
 		t.Fatalf("expected %v, got %v", want, repos)
+	}
+}
+
+func BenchmarkSearchRepositories(b *testing.B) {
+	n := 200 * 1000
+	repos := make([]*search.RepositoryRevisions, n)
+	for i := 0; i < n; i++ {
+		repo := &types.Repo{Name: api.RepoName("github.com/org/repo" + strconv.Itoa(i))}
+		repos[i] = &search.RepositoryRevisions{Repo: repo, Revs: []search.RevisionSpecifier{{}}}
+	}
+	q := "context.WithValue"
+	queryInfo, err := query.ProcessAndOr(q, query.ParserOptions{SearchType: query.SearchTypeLiteral, Globbing: false})
+	if err != nil {
+		b.Fatal(err)
+	}
+	options := &getPatternInfoOptions{}
+	textPatternInfo, err := getPatternInfo(queryInfo, options)
+	if err != nil {
+		b.Fatal(err)
+	}
+	tp := search.TextParameters{
+		PatternInfo: textPatternInfo,
+		RepoPromise: (&search.Promise{}).Resolve(repos),
+		Query:       queryInfo,
+	}
+	for i := 0; i < b.N; i++ {
+		_, _, err = searchRepositories(context.Background(), &tp, options.fileMatchLimit)
+		if err != nil {
+			b.Fatal(err)
+		}
 	}
 }