@stefanhengl 2020-10-08
> Would performance be better if we just divide the slice into 5 and avoid the cursor mutex? Could even just append all 5 found results at the end, and avoid all mutexes

Apparently not. The performance is statistically the same at least for higher levels of concurrency. Interestingly, the performance without the mutex seems to have a higher variance.

old: with cursor mutex
new: without cursor mutex

```
name                   old time/op  new time/op  delta
SearchRepositories      135ms ± 5%   142ms ± 4%  +5.81%  (p=0.000 n=10+9)
SearchRepositories-2   75.2ms ± 3%  77.4ms ± 2%  +2.93%  (p=0.007 n=10+10)
SearchRepositories-3   53.1ms ± 1%  55.1ms ± 4%  +3.78%  (p=0.000 n=8+9)
SearchRepositories-4   43.8ms ±11%  47.3ms ±12%  +7.92%  (p=0.016 n=10+11)
SearchRepositories-5   38.8ms ± 8%  41.3ms ±14%    ~     (p=0.061 n=10+11)
SearchRepositories-6   39.8ms ± 7%  40.8ms ±22%    ~     (p=0.863 n=10+11)
SearchRepositories-7   40.0ms ±13%  41.3ms ±10%    ~     (p=0.349 n=10+11)
SearchRepositories-8   39.5ms ± 6%  39.0ms ±17%    ~     (p=0.756 n=10+11)
SearchRepositories-9   38.1ms ± 8%  38.2ms ±11%    ~     (p=0.796 n=10+10)
SearchRepositories-10  38.9ms ± 8%  37.6ms ± 2%    ~     (p=0.133 n=10+9)
```