@keegancsmith 2020-10-08
LGTM. Can you share the diff for removing the mutex cursor? I'd be interested to micro optimize since I find that surprising :P