@stefanhengl 2020-10-08
```
diff --git a/cmd/frontend/graphqlbackend/search_repositories.go b/cmd/frontend/graphqlbackend/search_repositories.go
index a7bb985905..e8b685c9b4 100644
--- a/cmd/frontend/graphqlbackend/search_repositories.go
+++ b/cmd/frontend/graphqlbackend/search_repositories.go
@@ -103,19 +103,14 @@ func matchRepos(pattern *regexp.Regexp, resolved []*search.RepositoryRevisions)
 
 		muRepos = sync.Mutex{}
 		repos   []*search.RepositoryRevisions
-
-		muCursor = sync.Mutex{}
-		cursor   = 0
 	)
 	common.repos = make([]*types.Repo, len(resolved))
 	wg := sync.WaitGroup{}
-	worker := func() {
+	worker := func(start int, step int) {
 		defer wg.Done()
+		j := start - step
 		for {
-			muCursor.Lock()
-			j := cursor
-			cursor += 1
-			muCursor.Unlock()
+			j += step
 			if j >= len(resolved) {
 				return
 			}
@@ -152,7 +147,7 @@ func matchRepos(pattern *regexp.Regexp, resolved []*search.RepositoryRevisions)
 	*/
 	for w := 0; w < 5; w++ { // for benchmarking, replace 5 with runtime.GOMAXPROCS(0)
 		wg.Add(1)
-		go worker()
+		go worker(w, 5)
 	}
 	wg.Wait()
 	return common, repos
```