@keegancsmith 2020-10-08
Here is the full benchstat output with 5 runs each:

```
name                   old time/op  new time/op  delta
SearchRepositories      128ms ± 2%   135ms ± 2%   +5.18%  (p=0.008 n=5+5)
SearchRepositories-2   72.1ms ± 1%  69.3ms ± 1%   -3.80%  (p=0.016 n=4+5)
SearchRepositories-3   58.8ms ±10%  51.3ms ± 9%     ~     (p=0.056 n=5+5)
SearchRepositories-4   50.7ms ± 5%  48.2ms ±17%     ~     (p=0.841 n=5+5)
SearchRepositories-5   50.4ms ± 2%  43.9ms ± 7%  -12.90%  (p=0.016 n=4+5)
SearchRepositories-6   49.7ms ± 3%  42.2ms ±14%  -15.06%  (p=0.008 n=5+5)
SearchRepositories-7   48.3ms ± 3%  40.8ms ± 6%  -15.53%  (p=0.008 n=5+5)
SearchRepositories-8   47.8ms ± 2%  41.3ms ±12%  -13.60%  (p=0.008 n=5+5)
SearchRepositories-9   51.1ms ± 1%  40.6ms ± 5%  -20.53%  (p=0.008 n=5+5)
SearchRepositories-10  51.7ms ± 1%  40.6ms ± 2%  -21.42%  (p=0.008 n=5+5)
```